Rename peer list names and variables
