+++ elements/lisk-p2p/src/peer_directory/peer_book.ts
@@ -19,13 +19,13 @@ import {
 	DEFAULT_NEW_BUCKET_SIZE,
 	NewPeerConfig,
 	NewPeers,
-} from './new_peers';
+} from './new_list';
 import {
 	DEFAULT_TRIED_BUCKET_LENGTH,
 	DEFAULT_TRIED_BUCKET_SIZE,
 	TriedPeerConfig,
 	TriedPeers,
-} from './tried_peers';
+} from './tried_list';
 
 export interface PeerBookConfig {
 	readonly newPeerConfig?: NewPeerConfig;
@@ -74,41 +74,28 @@ export class PeerBook {
 	public getAllPeers(): ReadonlyArray<P2PPeerInfo> {
 		return [...this.newPeers, ...this.triedPeers];
 	}
-	/**
-	 * Description: When a peer is downgraded for some reasons then new/triedPeers will trigger their failedConnectionAction,
-	 * if the peer is deleted from newPeer that means the peer is completely deleted from the peer lists and need to inform the calling entity by returning true.
-	 */
-	public downgradePeer(peerInfo: P2PPeerInfo): boolean {
-		if (this._newPeers.getPeer(peerInfo)) {
-			if (this._newPeers.failedConnectionAction(peerInfo)) {
-				return true;
-			}
+
+	public getPeer(peerInfo: P2PPeerInfo): P2PPeerInfo | undefined {
+		const triedPeer = this._triedPeers.getPeer(peerInfo);
+		if (this._triedPeers.getPeer(peerInfo)) {
+			return triedPeer;
 		}
 
+		return this._newPeers.getPeer(peerInfo);
+	}
+
+	public updatePeer(peerInfo: P2PPeerInfo): boolean {
 		if (this._triedPeers.getPeer(peerInfo)) {
-			const failed = this._triedPeers.failedConnectionAction(peerInfo);
-			if (failed) {
-				this.addPeer(peerInfo);
-			}
+			return this._triedPeers.updatePeer(peerInfo as P2PDiscoveredPeerInfo);
+		}
+
+		if (this._newPeers.getPeer(peerInfo)) {
+			return this._newPeers.updatePeer(peerInfo);
 		}
 
 		return false;
 	}
-	// Move a peer from newPeer to triedPeer on events like on successful connection.
-	public upgradePeer(peerInfo: P2PPeerInfo): boolean {
-		if (this._triedPeers.getPeer(peerInfo)) {
-			return true;
-		}
 
-		if (this._newPeers.getPeer(peerInfo)) {
-			this._newPeers.removePeer(peerInfo);
-			this._triedPeers.addPeer(peerInfo as P2PDiscoveredPeerInfo);
-
-			return true;
-		}
-
-		return false;
-	}
 	// It will return evicted peer in the case a peer is removed from a peer list based on eviction strategy.
 	public addPeer(peerInfo: P2PPeerInfo): P2PPeerInfo | undefined {
 		if (
@@ -133,22 +120,38 @@ export class PeerBook {
 		return false;
 	}
 
-	public getPeer(peerInfo: P2PPeerInfo): P2PPeerInfo | undefined {
-		const triedPeer = this._triedPeers.getPeer(peerInfo);
+	// Move a peer from newPeer to triedPeer on events like on successful connection.
+	public upgradePeer(peerInfo: P2PPeerInfo): boolean {
 		if (this._triedPeers.getPeer(peerInfo)) {
-			return triedPeer;
-		}
-
-		return this._newPeers.getPeer(peerInfo);
-	}
-
-	public updatePeer(peerInfo: P2PPeerInfo): boolean {
-		if (this._triedPeers.getPeer(peerInfo)) {
-			return this._triedPeers.updatePeer(peerInfo as P2PDiscoveredPeerInfo);
+			return true;
 		}
 
 		if (this._newPeers.getPeer(peerInfo)) {
-			return this._newPeers.updatePeer(peerInfo);
+			this._newPeers.removePeer(peerInfo);
+			this._triedPeers.addPeer(peerInfo as P2PDiscoveredPeerInfo);
+
+			return true;
+		}
+
+		return false;
+	}
+
+	/**
+	 * Description: When a peer is downgraded for some reasons then new/triedPeers will trigger their failedConnectionAction,
+	 * if the peer is deleted from newPeer that means the peer is completely deleted from the peer lists and need to inform the calling entity by returning true.
+	 */
+	public downgradePeer(peerInfo: P2PPeerInfo): boolean {
+		if (this._newPeers.getPeer(peerInfo)) {
+			if (this._newPeers.failedConnectionAction(peerInfo)) {
+				return true;
+			}
+		}
+
+		if (this._triedPeers.getPeer(peerInfo)) {
+			const failed = this._triedPeers.failedConnectionAction(peerInfo);
+			if (failed) {
+				this.addPeer(peerInfo);
+			}
 		}
 
 		return false;
